+++ src/reinforce/__init__.py
@@ -32,7 +32,7 @@ from .algorithms import (
 )
 from .core import Box, Discrete, Env, Space, Transition, Wrapper
 from .data import TransitionDataset, collect_dataset
-from .registry import list_algorithms, list_environments, make_agent, make_env
+from .registry import list_algorithms, list_environments, make_agent, make_env, make_vec_env
 from .training import evaluate_policy
 from .utils import set_seed
 
@@ -80,6 +80,7 @@ __all__ = [
     "set_seed",
     "make_agent",
     "make_env",
+    "make_vec_env",
     "list_algorithms",
     "list_environments",
 ]